backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=13.0
google-genai>=1.55.0
google-auth>=2.35.0
pillow>=10.0.0
python-multipart>=0.0.12
pydantic>=2.9.0
httpx>=0.28.1
python-dotenv>=1.0.0
